+++ packages/engine/src/index.ts
@@ -9,3 +9,4 @@ export * from "./text/index";
 export * from "./sdk/index";
 export * from "./runtime/index";
 export * from "./preview/index";
+export * from "./export/index";